Clean up dynamic geometry updaters
The dynamic geometry updaters in the DataSource layer did not have
the same semantics as the rest of the visualizations system.  Mainly,
an undefined value would cause Cesium to crash instead of simply not
rendering an object.  This change uses the newer Property.get helper
functions in order to both clean up the code and follow the same rules.

Fixes #2252