Use dimensions, instead of min/max corners, for BoxGraphics
In order to be more consistent with other Entity geometries, such as
cylinder and ellipsoid, this changes BoxGraphics to use
`position`, `orientation`, and `dimensions` properties instead of
`minimumCorner` and `maximumCorner` properties.  This also allows for
rotated boxes, which the previous behavior did not.  This also opens up
the possibility for a dynamic box fast-path, similar to what we do for
ellipsoid, if the need ever arises.