Fix polygon geometry packing.
Even when `options.height` is undefined, we set it to 0.  We then compute
a separate `_extrude` boolean property based on `perPositionHeight`.  When we started packing/unpacking
geometry, the `_extrude` property would be incorrectly computed because
the `unpack` process would see `options.height` as 0 instead of undefined.

I did a quick review of the other geometry and it looked like polygons
were the only ones affected by this issue.  It wasn't caught by the packing
tests because the packing tests always caused `_extrude` to be false because
neither `height`, `extrudedHeight`, or `perPositionHeight` were set
in the test data.