Cache attributes returned by getGeometryInstanceAttributes.
Currently getGeometryInstanceAttributes generates a new object every time
it's called.  This generates unecessary garbage and forces the caller
to cache things himself in order to avoid it. This change simply makes it
so that the attributes object is cached after creation and returns the
same object the next time the function is called. This both reduces GC
pressure and is faster (since we aren't constantly rebuilding objects).