Don't set billboard.image values if they are the same.
Every time we assign a canvas to billboard.image it creates a new texture,
even if it's the same canvas instance. This is by design because canvases
can be drawn to multiple times. (like people trying to do billboard
animation). Since using a canvas causes this to happen every frame in the
BillboardVisualizer, we are constantly creating textures and adding them
to the atlas.

To fix this in the short term, I simply treat canvas elements as if it's
always the same image. This prevents the problem and the only
"feature" we lose is to keep drawing to a canvas and have it automatically
update in the billboard (which is probably the wrong way to do billboard
animation anyway since it's incredibly inneficient).

When I have some free time, I'll look into adding a proper billboard
animation demo to Sandcastle using the Entity API.

This also allowed me to remove some code from GeoJSONDataSource.