Workaround issue in Firefox 35
In Firefox 35 we need to force the context alpha to true because setting
it to false affects all framebuffers. Fixes #2431

Also removed ancient workaround for a Chrome issue that was fixed in
Feb 2014 as well as an unused glsl #ifdef that was left over from an
older version of Firefox.