Fix OIT on machines where it was broken.
As pointed out by one of the Chromium developers, we were using an undefined
variable in `czm_alphaWeight` due to a double declaration of the `z` variable
resuling in scoping issues.  This change then lead to another issue that
where textured were black, which was fixed by flipping `q.w`

We also discovered that #2327 is related to `czm_alphaWeight` and may
be an issue with the weighting function. @bagnell is going to look into it separately.