Model.boundingSphere was not properly scaling its center.
When a scale is applied to a Model primitive, it properly scales
the radius of `Model.boundingSphere` but it was not scaling the center
to match.  This is easy to see by using the Sandcastle 3D Models example
and setting a `scale : 10` on the models and then switching to Cesium Man.
Instead of viewing the center of the model, as you get with scale of 1,
you end up looking at his feet.

Also added a test to check for this, since we didn't have one.  The test
passed in this branch but fails in master.