Fix an imagery loading bug.
When a terrain tile maps to a non-level-zero tile, and that non-level-zero imagery fails to load, the terrain tile will never be considered done loading.  If the imagery tile must be renderable before the terrain tile is renderable, the terrain tile will never be renderable and part of the globe will be missing.

The fix is for TileImagerys to process the statemachine of not just the imagery tile they'd like, but also the ancestor imagery tile they're stuck with.  Previously we relied on other terrain tiles to move the imagery tile through the load process, but this fails in the case described above.