Fixes a bad test.
RectanglePrimitive's 'renders bounding volume with debugShowBoundingVolume' was attempting to assert that the bounding volume was rendered.  But the `render` function used in our specs doesn't know how to render bounding volumes.  The only reason this passed is because previous tests were drawing something and the buffer wasn't being cleared.  If this test was run by itself or if the tests were run in a different order, this test would fail.

This changes the test to just check for the presence of the flag on the command, rather than rendering anything.