Fix infinite loop in drillPick
As pointed out by @mantelpiece in #2442, if batched Primitives do
not have a show attribute, we end up in an infinite loop because
the primitive is never hid.

This change fixes that problem as well as cleans up drillPick in general
to be a little faster and easier to read.  Also added tests for batched
primitive picking, of which we had none.