Improve jasmine's handling of possibly-async functions.
Jasmine doesn't like it when an async spec (one that takes a done parameter) actually ends up being synchronous and calls `done()` before the spec returns.  It mostly works, but ends up creating a deep call stack that risks causing a stach overflow.

This change makes jasmine detect when done() is called before the spec returns and treat the spec as it would a normal synchronous spec.  This should improve performance a bit too because all Cesium specs are async from the perspective of Jasmine, but only some of them actually are.