Fix changing static point colors
As [reported on the forum](https://groups.google.com/d/msg/cesium-dev/plbxGWaPxr8/twTdRAzbC8UJ),
there was a false optimization that would cause changing a point's color
from one constnat value to another to have no affect.

I fixed the issue and update an existing test (which was old and partially
incorrect anyway).  The test fails in master bu passes in this branch.