Work around inconsistent browser behavior.
Firefox does not allow you to add elements to an iFrame until the load
event is fired, even if you are building everything manually.
Chrome does not send the load event unless we explicitly set a src.
Solution is to always used a load event and set the src to about:blank.

Also cleaned up the InfoBox constructor and expose the frame element
we create directly.