Fix perVertex colors for both PolylineGeometry and SimplePolylineGeometry
This was due to a typo in the code (and no matching unit test for it). We
were using `perVertex` internally and `colorsPerVertex` for the API. In
addition to fixing the typo, I changed it to `colorsPerVertex` to avoid
confusion in the future.  I also updated tests to make sure everything
was what it should be.