Fix InfoBox crash
`window.getComputedStyle` can return `null`, which I didn't realize,
leading to a crash when assigning unstyled text.

I also improved the specs for this (which were non-existent).  The new
specs fail in master but succeed in this branch.

Finally, I fixed a stupid KML typo that was specifying `san-serif` instead of `sans-serif`