Fix InfoBox on Mac
On Chrome and Firefox in Windows, `window.getComputedStyle` works on any
node and returns null for non-element nodes.  Under the same browsers on
OS X, `getComputedStyle` throws for non-element ndoes, leading to test
failures and a crash in the Cesium WMS picking demo.

This change simply verifies that the first child is indeed an element
before calling `getComputedStyle`.