Improve GeoJSON tolerance for null values.
It's legal to include `null` values in many places with GeoJSON but we
weren't always handling these cases in Cesium.  This change introduces a
new helper, `definedNotNull`, which as the name implies checks for both
`undefined` and `null`.  We now use this in GeoJsonDataSource when reading
parsed JSON. We treat `null` in JSON the same as undefined, unless the
specification explicitly says otherwsie.  One exception to this is
`desription` which I interpret null as meaning "this object has no
description" instead of us generating a default.