Asynchronous Primitive not setting ready on Failure
For usability, we allow users to create "invalid" geometries, such as
a polygon with two points, and simply don't render anything as a result.
Unfortunately, this case was not being handled for unbatched asynchronous
geometries, which resulted in the Primitive never being ready.

This change fixes that issue and also adds tests for both synchronous and
asynchronous cases (the sychronous case was already being handled correctly
but was untested).

Fixes #2591