Fix crash when adding and removing a billboard before next update.
As [reported on the forum](https://groups.google.com/d/msg/cesium-dev/Ti4jLsugBq4/13Qo5_daa68J),
adding and then removing a billboard before the collection is updated
causes a crash because we don't remove the billboards before trying to
recreate the atlas.  The new test I added fails in master but passes in
this branch.