Don't call initializeFrame when the canvas size is 0
Having the default render loop call `scene.initializeFrame` when the canvas
height or width is 0 causes issues in camera code which expects these
values to be non-zero.  As far as I can tell, similar to `render`, there's
no reason we should be calling `initializeFrame` in this case anyway.

Fixes #2662 (some additional discussion there as well)