Fix crash on entity double click.
When an entity geometry exists forever but its position property is
time-limited.  Double clicking on the geometry would attempt to
track that position, which is undefined at the current animation time.
We now better guard against defined Property instances with undefined
values and instead zoom to the extent of the geometry.