Don't complete an entity zoom while morphing.
If viewer.zoomTo found itself in the middle of a morph it would crash. We
now wait for the morph to complete.  I couldn't think of a good unit for
this because it's a race condition, but there are instructions in the linked
issue on reproducing it in Sandcastle.  Fixes #2680