Fix material caching for canvases
Cesium would incorrectly use an HTMLCanvasElement as the key to the Material
texture cache dictionary.  This means that if you had multiple canvas-based
materials, we would only ever use the first one that was added (because
the cache lookup would always succeed even if the canvases were difference).

This change uses toDataURL to store canvas elements in the cache.  This
is sub-optimal since toDataURL is slow and also creates huge string,
but the best we can do in the short term.  We need a good overlal long
term strategy for dealing with canvases, but that's outside the scope of
this PR.