Remove Material texture cache
The Material texture cache caused many more problems than it solved, so
this removes it until a complete solution can be found.

This fix exposed several bad tests that were actually using the default
texture instead of waiting for an image material uniform to load.  These
tests have been fixed.

This fix also exposed another bug as well, which caused a material
rendered with the default texture and then destroyed to end up destroying
the default texture.  Since textures are not reference counted, I simply
added an `isDefault` flag for now and had Material.destroy only destroy
the texture if it was not the default texture.