Make most tests pass on IE11
Apparently the IE 11 test failures we've been seeing for months were
actually a bug in Cesium where we try to destroy `scene._globeDepth`
even when it was undefined. Fixing that allowed IE11 to once again
run tests to compeltion without thousands of errors. This then revealed
several other spec-only issues that needed to be fixed as well.

This brings the overall number of failures down to 5, all of which look
rendering related.  I'll write up a separate issue for them.