add ArcGIS MapServer token based security support
This adds support for ArcGIS's proprietary token-based authentication for
services. [1]

The solution is not robust: tokens are subject to expiration, after which point
the imagery provider will begin throwing exceptions (rightfully so).

Abstracting authentication further (outside of ArcGIS MapServer imagery
providers) may be a good idea, but is a larger task worth more thought.

Further, ArcGIS supports requesting security tokens for a set of credentials
via HTTP (GET, possibly POST as well), though this is not guaranteed to be
enabled on the server.

[1] - http://resources.arcgis.com/en/help/main/10.1/index.html#/About_ArcGIS_tokens/0155000005s0000000/