Also bind tilt to ctrl-right-drag.
On Firefox, on OS X, it is impossible to ctrl-left-click, on a laptop with a trackpad.  Because OS X typically interprets ctrl-click as a secondary ("right") click, Firefox "helpfully" decides to send a right-click JS mouse event, unlike both Chrome and Safari.

This new mouse binding affects all platforms, but ctrl-right-drag wasn't bound to anything by default anyway.