Memory tweaks
1. Have `AssociativeArray.remove` actually delete the key from the object
hash instead of just setting it to undefined, avoiding unbounded object
growth.

2. If Entity geometry was removed and a geometry of the same type was
never re-added back, the destroyed Primitive instance would remain in a
closure, causing any JavaScript objects referenced by the Primitive to hang
around, in the case of entities, this was the array of instance ids, which
could be fairly large.

3. In order to reduce the change of 2 being a problem in other cases, I
marked additional properties as undefined in `Primitive.destroy` to avoid
objects hanging around after it is called even if the primitive itself
is still inside of a closure.