Set shuttle ring to max angle when multiplier is too large.
Previously, when the clock had a multiplier larger than the largest
shuttle ring tick mark, the dial marker would end up going offscreen making
it impossible to adjust the multiplier through the UI.  Ancient code
tried to account for this by expanding the list of valid ticks to include
the new multiplier.  Unfortunately, that approach isn't ideal because it
would change user settings out from under them.  Instead, this approach
simply sets the dial to it's max value when the muliplier is larger
and the first click to slow things down will snap to the largest
actual tick.

Fixes #1072