Be more lenient when processing KML coordinates
Our code in master is pretty strict whenprocessing KML, but it turns
out Google will let you pass almost anything and just use zeros (in the
case of single coordinates) or create an object without an geometry (in
the case of linear rings).  This change makes Cesium match that behavior,
event though the spec itself says it's invalid.

I also noticed the Viewer specs had a bug because Viewer itself was listed
in the list of requires (the way our specs work, the thing being tested
has to be in the list of requires and has to be first).  In this case,
because all creation is deferred to the `createViewer` helper, it wasn't
being included.

To test this just run/review the unit tests, they are updated to reflect
the change in functionality.