Fix TestWorkers in TaskProcessorSpec.
Recent changes (well, over a year ago) to RequireJS broke the way we were locating the Source directory in both normal and combined test modes. Instead, we now configure RequireJS with an additional path "Source" which always points to the Source directory, regardless of mode.

This path should not be used from other specs, because it will bypass the stubs in the combined test mode. It is specifically intended for the TaskProcessor specs, because the TestWorkers need to always load using the individual modules in Source.