Fix when.defer() usage
It turns out that there are a bunch of places in our code where we
use `when.defer()` and return the deferred instance instead of its
`promise` property.  This works in most cases because when's deferred
promises are thenable, however they don't have all of when's API, like
`always`.  This change makes sure we correctly return `deferred.promise`
everywhere.

It's probably a quirk of the version of when we are using
that defers aren't always-able and always isn't part of the A+ spec,
so there's no reason to add additional tests for these.