Adds the 'selectable' member to the Entity class
Entity now has a 'selectable' member which determines whether the given
Entity is selectable by the user.  'selectable' defaults to true, and can
be set via the 'selectable' field in the options parameter of the Entity
constructor.  Additionally, CzmlDataSource has been updated to handle the
'selectable' field and set an Entities 'selectable' value.