Fix Eclipse jsHint integration
Because jshint-eclipse does not support multiple `.jshintrc` files, I've
updated the root .jshintrc to include jasmine.  This is then set to
false in lower-down `.jshintrc files using the extends capability.

This has the added benefit of meaning we no longer write out .jshintrc
files as part of the build step.

The only downside here is that jsHint in Eclipse thinks Jasmine globals
are okay in all of our Source files. However, that's not a big deal
and travis and our own jshint build step will detect the problem if
someone tries to call jasmine outside of the Specs directory.