Add support for video materials
This extends the `Material` object to support `HTMLVideoElement`, similar
to how we already support `Canvas`. It also adds similar support to the
`Entity` layer, via the `ImageMaterialProperty`.

Also added `VideoSynchronizer`. By default, video materials just use the
current video frame as the texture, so if you pause, seek, or otherwise
change the video, the material is updated to match. `VideoSynchronizer` is
a helper class that flips this so that the video time is controlled by the
simulation clock, allowing for you to synchronize the video with the rest
of your scene.

Added a new Sandcastle example using Big Buck Bunny, an open source blender
movie project from a few years ago.  To avoid bloating the repository and
release zip size, movie files are hosted on cesiumjs.org.

Browser Support
IE will most likely never support video textures, but they are being
added to Edge. Chrome and Firefox work great, Safari should work as well.

Sound
Sound works as expected, even when playing faster or slower than normal. Ultimately it's browser controlled and users determine whether to include sound or not directly on the video element. Our Sandcastle demo has sound.