Remove deprecated synchronous XHR request from Sandcastle
Synchronous XHRs are deprecated and they were never really a good
practice to begin with.  This updates Sandcastle to remove synchronous
requests when loading demos.

This also makes loading Sandcastle technically faster, but loading right
now is deprecated by the use of Cesium in AMD form (which we should also fix).