Fix incorrect dojo all usage in Sandcastle.
All returns a thenable and does not take a then action as a second
parameter.  This was preventing the all button from ever being added.

There was also a separate issue that would cause the first loaded example
end up last in the display list. This is because we were waiting on
promise resolve for the loading for no good reason.

Fixes #3266