Fix Sandcastle jsHint
We were writing out jsHintOptions with the `extends` attribute, which
wasn't working when loaded into the browser.  Now the build process
loads the primary jshint file and modifies is to match the gallery
in order to write out a complete jsHintOptions object.

This could probably be a little more robust (since not all changes
to Apps/Sandcastle/.jshintrc get reflected automatically jsHintOptions,
but I doubt that will be an issue given that these files will rarely change.