Honor devicePixelRatio take 3
This reintroduces #3233, but now uses the `image-rendering` CSS property
to disable smooth scaling and avoid burriness. Since this is a CSS4
property, it's not yet supported by all browsers (Safari/IE), so we detect
if the property is supported and if it isn't, we fallback to full
resolution rendering on those browsers so that the image is always crisp.

The other nice thing about this approach is that it's detect capability
and not the browser itself, so when Safari and IE do add support, Cesium
will automatically start taking advantage of it.