Make Cesium work with current release of gdal2tiles.py
gdal2tiles.py previously had 2 bugs compared to the TMS specification.

1. Origin and BoundingBox flipped X and Y values.
2. Origin and Bounding box were always in geodetic degrees,
regardless of the file.

They fixed 1, but 2 is still an issue. Our workarounds were kind of smashed
together in the code so I detangled it to remove doing 1 by default, but
added a `flipXY` property for manually enabled backwards compatibility.
Unfortunately, the buggy and non buggy versions of gdal generated identical
files except for the flipped values, so it's impossible for us to detect
this ourselves.

I also fixed bad promise resolution in tilemapresource.xml, which would
silently fail (in the original code) when `flipXY` was performed but no
longer needed. Fixing this makes some buggy TMS sets load properly anyway,
so that's a win for us.

For TMS layers we have in the repository, I fixed them to not contain the
bug.  Only tests that specifically check for `flipXY` behavior include
the buggy XML.

I did NOT change the black marble data and instead added `flipXY` to the
Sandcastle example and defaultLayers code.  We can certainly regenerate it
if we want, but it will break old versions of Cesium that use it.