Fix travis timeout
Apparently #3430 is still an issue and the travis_wait command added as
part of #3432 does not work as advertised. To address the issue once and
for all, I turned up the loglevel on requirejs so that it always outputs
progress to the build log (this was actually requested by @shunter and
a few others after the logging went away when we moved to gulp anyway).

Long story short we should finally stop seeing travis timeouts.

I also make jsHint run as a separate step and instead of `release` we
run `makeZipFile` to be sure the entire build process works.  If
`makeZipFile` were to error, the previous configuration wouldn't have
caught it.