Fix gzip compression settings.
The server was blindly adding `Content-Encoding=gzip` too all requests for
b3dm files. When the request was for a non-existent tile, the 404 message
would also be marked as gzip, causing the browser to fail decoding it
(different browsers have different behavior in this case).

The correct solution is to use the `setHeaders` callback as part of express
static serving, which will only get called to set the headers when the
file actually exists.