Deploy zip and npm packages
This change allows travis to deploy the generate zip files and
npm packages in addition to the build output.  Each package
gets a unique name based on branch name and build number.
Also updated deployment step not to delete these packages so that
they can be depended on until the branch is merged.