Add ability to create open extruded PolygonGeometry
While working on the NYC 3D Tiles demo, I needed to create open Polygon
geometry (for example, a building with a different color roof). Having a
top and bottom on the polygon was inefficient and also caused z-fighting.)

Also added feature to CZML and Entity API as well as tweaked Sandcastle
examples to show it.