Fix GlobeSurfaceTileSpecs
As part of #3900, `vertexArraysToDestroy` was added as a parameter to
`GlobeSurfaceTile.processStateMachine`, but none of the tests actually
pass that parameter.  This happens to be OK in master because none of
the tests end up using the parameter, but in the 3D Tiles branch, the
parameter is required and merging in master results in failing tests.

Passing in an empty array fixes the issue (since we don't care about
destroying the data in the specs), but we might possibly want to actually
test that the parameter is being used correctly.