Do not attempt to deploy fork pull requests to s3.
For a pull request, the `TRAVIS_BRANCH` variable is always `master` this
causes deployment to deploy to the master branch instead of the branch
name (which is unavailable for pull requests). Even if the branch name
were available, forks that open pull requests could have name conflicts
with branches in the primary repository, so we wouldn't be able to deploy
them anyway.

Fixes #3937