Fix viewer.zoomTo/flyTo for certain arguments.
ES6 (Chrome 51 and Firefox 48) now expose an `Array.values` function.
This broke existing code in Viewer that was relying on duck-typing
to detect if an object was an EntityCollection or not.  Added logic to
detect if the zoomTarget was already an array for early return.

Fixes #3980, see that issue for more details.