Move logic for real-time mode (SYSTEM_CLOCK) into Clock itself.
SYSTEM_CLOCK implies animating forward at a 1x multiplier.  Centralize the logic for dealing with coordinating clock settings in Clock itself.  When switching to SYSTEM_CLOCK, other values are immediately updated to be consistent.  When changing other values, e.g. current time or multiplier, SYSTEM_CLOCK is deactivated.

This fixes some inconsistencies in the UI where the logic was not consistent, due to being scattered around.  Specifically, when you dragged the timeline while in realtime mode, you ended up in a hybrid state where the mode was still set to SYSTEM_CLOCK and therefore the label in the UI still said "Today" etc.