Fix handling of pointermove on touch devices
`ScreenSpaceEventHandler` did not handle the case where a `pointerdown`
happens on an element other than the canvas and then use held down and
moved onto the canvas.  Essentially, it always assumed that the list of
positions in the initial pointerdown were available in the pointermove,
they aren't.

This problem only happens on touch devices with PointerEvent.