Fix ground primtives with dynamic color
Ground primitives with dynamic color exposed a bug in
`StaticGroundGeometryColorBatch` where we were not removing ground
primitives if the batch size went to 0.  Fixing that bug exposed another
bug which is the fact that dynamic colors for ground primitives need to be
treated as dynamic geometry beause of the lack of per instance coloring
for batched ground primitives.

Fixes #4433