Further improve clamping
Previously, we would only clamp a billboard or label at it's highest
terrain level, we wouldn't "go backwards". This was thought to be
a performance benefit.  However, in practice it makes it really easy
for clamping to be incorrect and you end up with billboards either
constantly floating or below ground.  This change removes that further
optimizaiton and as far as I can tell, clamping is now always correct.

The downside of this is that we are doing a big more work and clamping
was already slow.  However, since most of the time clamping is doing
nothing and then bursts, upping the time slice improved performance without
any major degradation in frame rate. Once we further optimize the terrain
picking code (which is why clamping is so slow to begin with), we could
probably further reduce the timeslice.