Timeline should not use ISO8601 as an intermediate format.
This fixes #4544 and replaces #4553.

Basically, makeTics was using ISO8601 and JavaScript dates as an
intermediate format for computing it's Julian epoch.  This lead to all
kinds of problems for very early or late dates.

This fixes the crash and everything works, but there are still labeling
problems when you go before 4713 BCE, but I'll write up a new issue to
deal with that separately since it's a mode widespread Cesium problem.