Billboard halfsize must be computed after compressedAttributes.
Compressed vertex attributes must be integers.  Billboard halfSize is
needed for computing the distance from the center of the billboard out
to one of the corners.  But the half-size can have a .5 fractional
component, from the least significant bit of the original billboard
size.  So this 0.5 must be computed after the compressed vertex
attribute has already been unpacked.