Fixed a crash when using Entity path visualization with reference properties.
Fixes #4915

There were several bugs here:

1. `PathVisualizer` was unwrapping a `ReferenceProperty`, finding a
`CompositePositionProperty` inside, and then using the reference as if
it was a composite instead of using the inner property.
2. Same was 1, but for `TimeIntervalCollectionPositionProperty`
3. We were unwrapping `ScaledPositionProperty` and using the inner property
directly, but this prevented the value from actually being scaled properly.
Now we use the `ScaledPositionProperty` which means it falls back to a
generic sampling path but that's the best solution. (Not to mention that
ScaledPositionProperty is privat and will eventually go away).