Fixed crash in Camera.pickEllipsoid when height or width is 0
There are cases where we programatically call `pickEllipsoid`, usually
with the center or corner points of the screen. When the canvas is
currently hidden via CSS or hasn't been added to the DOM yet, the canvas
has a length and width of 0.  Calling pickEllipsoid in these cases would
lead to NaNs during Cartesian3.normalize.  The correct behavior is to
simply return undefined, as we do in other cases where pickEllipsoid is
not possible or fails.