Support Animation widget construction outside of the DOM
Creating a new `Animation` instance would crash if provided container was
not currently in the DOM. This is because we use computed styles for
themeing and computed styles return undefined if the element isn't in the
DOM.  Since we call `applyThemeChanges` during construction, this would
lead to a crash.

The solution is check if the container is in the DOM, and if not,
use `MutationObserver` to delay the style application until the element
is added.