Viewer clock improvements
In master, it's impossible specify the `clockViewModel` instance to use
for any child widgets that need it.  Viewer also doesn't expose the
`ClockViewModel` and you have to dig into
`animation.viewModel.clockViewModel` to get to it.  This makes it hard to
write custom widgets that need to share the same `ClockViewModel` and are
not directly tied to the lifetime of the `clockViewModel` created by
Viewer (which gets destroyed when the Viewer is destroyed).

This change deprecates the Viewer `clock` constructor option and replaces
it with `clockViewModel`.  If a custom `clockViewModel` is specified, it
is passed through to any child widgets that need it and it is not destroyed
on destruction.

We didn't have tests for the old functionality, but I added a test for
the new parameter.