Fix KML descriptions that link to relative paths
If a KML description included a relative path, the browser request would
end up using whatever server the application was hosted on as the base
path, this is because the KML HTML would be injected into the InfoBox
as-is. The end result is that relative images and links inside of KML
files would not load properly.

To fix this, we now expand all relative paths into absolute paths that
use the KML's source Uri as the bas.