Fix sporadic build/test failures.
The default load timeout for AMD modules is 7 seconds, when running the
tests so many modules get loaded at once that it can time out. This fixes
the karma configuration under Node so that timeout is indefinite.

Because we've also seen sporadic build failures, I also added the timeout
option to `requirejs.optimize` as well, I'm not sure if this has any
effect, but if we stop seeing sporadic requirejs.optimize failures, we'll
know it did.