Fix travis builds
The problem was that travis itself was killing our build process because
we were trying to run 8 requirejs processes at once, leading us to run out
of memory (only 3GB available on travis).  This adds a command line to
specify concurrency and updates travis to do at most 2 (one for each core).