Return null when adding a completed tween
Previously, when trying to add a tween with a duration of 0, its complete
function was immediately called, it was not added to the collection and
a fake/empty tween was returned.

This was bringing confusion in the flyTo code where the result of the add
is put into the current_flight variable.

With this change a null is now returned.