Fix Cesium ion external asset handling
`CesiumIonResource` should only be used for ion hosted assets, so create
a standard Resource instance for external 3D Tiles and Terrain.  This also
means that `CesiumIonResource.createImageryProvider` can't exist and needs
to just be on `CesiumIon` directly instead.

Also add `Instrumented` directory to .eslintignore becuase it was linting
the results of `instrumentForCoverage`