Cesium ion credit handling
1. Remove `CesiumIon.create3DTileset`, it is no longer needed now that
`Cesium3DTileset` can take a promise as the `url` parameter.  Also update
`Cesium3DTileset` to show credits for ion assets.
2.Remove `CesiumIon.createImageryProvider` and replace it with
`CesiumIonImageryProvider` this was needed to show credits for ion assets
in all cases, but is also more consistent with the rest of our API.
3. Update `CesiumTerrainProvider` to show credits for ion assets.
4. Move `CesiumIon` and `CesiumIonResource` to `Core`
5. Minor `Credit` clean-up by @hpinkos
6. Move default Cesium Logo credit from `CesiumWidget` to the
`CreditDisplay` itself. This means the logo now shows in all cases. When
ion assets are used, we swap the logos to avoid cluttering the display.
If a user removes the Cesium log, the ion logo will still be added since
it is a required attribution.