Allow Viewer imageryProvider or terrainProvider with BaseLayerPicker
We've had a check in Viewer forever to see if the baseLayer picker was
being used when a terrainProvider or imageryProvider was specified.  This
lead to a tone of code where we create the viewer and then just specify
viewer.terrainProvider anyway. Basically, this check was pointless and
just made the code more verbose.  The check also prevented devs from
specifing a terrainProvider but still using BaseLayerPicker for imagery
(and vice versa). Now we just let the user do what they want, which was
probably the right call from the beginning.