STK World Terrain -> Cesium World Terrain
1. Remove all `STK World Terrain` usage/references and replace them with
`Cesium World Terrain`.
2. Add new `createWorldTerrain` helper function for creating the world
terrain. Same defaults for extensions as CesiumTerrainProvider(no normals
or water).
3. Replace all World Terrain examples and doc with `createWorldTerrain`.
We were needlessly setting `requestWaterMask` and `requestVertexNormals`
to true in most examples, so I just changed it to use the default behavior
to simplify the code.
4. Add a deprecation warning and deprecation Credit when `assets.agi.com`
is detected. Note that this includes ArcticDEM and PAMap, which are still
being used in this PR but will be replaced in a follow-up PR after this
is merged.